fix(tests): move NotificationsPage and InteractionsPage to components dir
NotificationsPage is a global overlay utility (portal-rendered
notifications) that can't be scoped to a single container. The body
container scoping was causing test failures across multiple shards.
InteractionsPage is a page.mouse utility with no meaningful container.
Both moved to components/ (exempt from scope-lockdown) with original
this.page locators restored.
