+++ packages/testing/playwright/pages/components/InteractionsPage.ts
@@ -1,15 +1,11 @@
 import type { Locator, Page } from '@playwright/test';
 import { expect } from '@playwright/test';
 
-import { BasePage } from './BasePage';
+import { BasePage } from '../BasePage';
 
 export class InteractionsPage extends BasePage {
 	constructor(page: Page) {
 		super(page);
-	}
-
-	get container(): Locator {
-		return this.page.locator('body');
 	}
 
 	async precisionDragToTarget(
